requirements.txt
aiofiles==23.1.0
aiogram==3.0.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.3
fake-useragent==1.2.1
frozenlist==1.4.0
h11==0.14.0
idna==3.4
magic-filter==1.0.11
multidict==6.0.4
outcome==1.2.0
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
requests==2.31.0
selenium==4.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.7.1
undetected-chromedriver==3.5.3
urllib3==2.0.5
websockets==11.0.3
wsproto==1.2.0
yarl==1.9.2
